Rename utils module to std_ext with inner modules
